handle relative links to "filename" or "./filename" correctly
Removed the test with File.expand_path(to,target_file),
it was always wrong with to = "filename" or "./filename".
The correponding unit test has been removed as well.

For hardlinks, load_current_resource sets @current_resource.to to ""
when the link doesn't exist or when it doesn't match.
The link is simply created when @current_resource == "".
The unit test has been modified accordingly.